BLOGGY app variable changed to application and renamed bloogy dir to app
