+++ app/models.py
@@ -1,4 +1,4 @@
-from bloggy import db
+from app import db
 
 class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
