+++ app/routes.py
@@ -1,8 +1,9 @@
-from app import app
-from flask import render_template
+from app import application
+from flask import render_template, flash, redirect, url_for
+from app.forms import LoginForm
 
-@app.route('/')
-@app.route('/index')
+@application.route('/')
+@application.route('/index')
 def index():
     user = {'username': 'Miguel'}
     posts = [
@@ -21,3 +22,11 @@ def index():
     ]
     return render_template('index.html', title='Home', user=user, posts=posts)
 
+@application.route('/login', methods=['GET', 'POST'])
+def login():
+    form = LoginForm()
+    if form.validate_on_submit():
+        flash('Login requested for user {}, remember_me={}'.format(
+            form.username.data, form.remember_me.data))
+        return redirect(url_for('index'))
+    return render_template('login.html', title='Sign In', form=form)
