--- bloggy/requirements.txt
@@ -1,31 +1,0 @@
-alembic==1.8.1
-attrs==22.1.0
-click==8.1.3
-distlib==0.3.5
-filelock==3.7.1
-Flask==2.2.2
-Flask-Migrate==3.1.0
-Flask-SQLAlchemy==2.5.1
-Flask-WTF==1.0.1
-importlib-metadata==4.12.0
-importlib-resources==5.9.0
-iniconfig==1.1.1
-itsdangerous==2.1.2
-Jinja2==3.1.2
-Mako==1.2.2
-MarkupSafe==2.1.1
-packaging==21.3
-platformdirs==2.5.2
-pluggy==1.0.0
-py==1.11.0
-pyparsing==3.0.9
-pytest==7.1.3
-pytest-assume==2.4.3
-six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
-SQLAlchemy==1.4.41
-tomli==2.0.1
-virtualenv==20.15.1
-Werkzeug==2.2.2
-WTForms==3.0.1
-zipp==3.8.1
-numpy
