--- bloggy/templates/index.html
@@ -1,9 +1,0 @@
-
-{% extends "base.html" %}
-
-{% block content %}
-    <h1>Hi, {{ user.username }}!</h1>
-    {% for post in posts %}
-    <div><p>{{ post.author.username }} says: <b>{{ post.body }}</b></p></div>
-    {% endfor %}
-{% endblock %}
