+++ microblog.py
@@ -1,1 +1,1 @@
-from app import app
+from app import application
